requirements.txt
matplotlib
pyserial
pyqtgraph
PyQt5
psutil